native.bbclass: poison TARGET_CFLAGS etc for native builds
TARGET_* FLAGS are never passed to the native toolchain. However,
they are exported so their values affect the checksum of every task.

Replace with fixed values to both avoid any unnecessary rebuilding
of native packages when target flags are changed (as before) and now
also to poison any accidental use of target flags with the native
toolchain.

Change-Id: Iae586daf5a58293ab6c168b7343bf814b1338da6
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>